Tuvn

+++ routes/user.js
@@ -254,15 +254,16 @@ router.post('/upload/:id', async function (req, res, next) {
         });
       }
 
-
-
       let results = req.files.map(async (file) => {
         const user = await userModel.findOne({ _id: _id });
         var filePath = user.avatar;
 
-        if (fs.existsSync('./public/'+filePath)) {
-          fs.unlinkSync('./public/'+filePath);
+        if (filePath) {
+          if (fs.existsSync('./public/'+filePath)) {
+            fs.unlinkSync('./public/'+filePath);
+          }
         }
+
         const entry = await userModel.findByIdAndUpdate({ _id: _id }, {
           avatar: `upload/users/${file.filename}`,
           modified: {
